predict/requirements.txt
mysql-connector-python==9.1.0
pandas==2.2.3
torch==2.5.1
numpy>=1.26.0,<2.1.0
requests==2.32.3
scikit-learn==1.5.2
python-dotenv==1.0.1
tensorflow==2.18.0
geopy==2.4.1
sentry-sdk
